requirement.txt
langchain==0.0.281
PyPDF2==3.0.1
python-dotenv==1.0.0